final-endterm/api/requirements.txt
fastapi
sqlalchemy
pyjwt
psycopg2-binary
uvicorn